Refactor imports to use sharedInterfaces
