+++ src/reactviews/pages/ExecutionPlan/executionPlanView.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as ep from "./executionPlanInterfaces";
+import * as ep from "../../../sharedInterfaces/executionPlanInterfaces";
 
 export class ExecutionPlanView {
     private _diagram: any;
