+++ src/reactviews/pages/ExecutionPlan/properties.tsx
@@ -5,7 +5,7 @@
 
 import "./executionPlan.css";
 
-import * as ep from "./executionPlanInterfaces";
+import * as ep from "../../../sharedInterfaces/executionPlanInterfaces";
 import * as utils from "./queryPlanSetup";
 
 import {
