+++ src/reactviews/pages/QueryResult/queryResultUtils.ts
@@ -3,9 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ColorThemeKind } from "../../common/vscodeWebviewProvider";
 import { QueryResultWebviewState } from "../../../sharedInterfaces/queryResult";
 import * as qr from "../../../sharedInterfaces/queryResult";
+import { ColorThemeKind } from "../../../sharedInterfaces/webview";
 
 export const saveAsCsvIcon = (theme: ColorThemeKind) => {
     return theme === ColorThemeKind.Light
