+++ src/reactviews/pages/TableDesigner/tableDesignerStateProvider.tsx
@@ -4,9 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as designer from "../../../sharedInterfaces/tableDesigner";
+import { WebviewContextProps } from "../../../sharedInterfaces/webview";
 import { getCoreRPCs } from "../../common/utils";
 
-import { useVscodeWebview, WebviewContextProps } from "../../common/vscodeWebviewProvider";
+import { useVscodeWebview } from "../../common/vscodeWebviewProvider";
 import { ReactNode, createContext, useRef, useState } from "react";
 
 export interface TableDesignerContextProps
