+++ src/sharedInterfaces/form.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { WebviewContextProps } from "../vscodeWebviewProvider";
+import { WebviewContextProps } from "./webview";
 
 export interface FormState<
     TForm,
